+++ client/src/App.jsx
@@ -1,8 +1,8 @@
 import "./App.css";
 import DashboardPage from "./pages/dashboard";
 // import LoginPage from "./pages/login";
-// import CreatePage from "./pages/create-user";
-// import EditPage from "./pages/edit-user";
+// import CreatePage from "./pages/create-user/create-user";
+// import EditPage from "./pages/edit-user/edit-user";
 
 function App() {
   return (
