+++ client/src/pages/create-user/create-user.jsx
@@ -1,0 +1,7 @@
+import CreateComponent from "../../components/create-user-component/create-user-component";
+
+function Create() {
+  return <CreateComponent />;
+}
+
+export default Create;
