--- client/src/pages/create-user/index.jsx
@@ -1,7 +1,0 @@
-import CreateComponent from "../../components/create-user-component";
-
-function Create() {
-  return <CreateComponent />;
-}
-
-export default Create;
