+++ client/src/pages/edit-user/edit-user.jsx
@@ -1,0 +1,7 @@
+import EditComponent from "../../components/edit-user-component/edit-user-component";
+
+function Edit() {
+  return <EditComponent />;
+}
+
+export default Edit;
