--- client/src/pages/edit-user/index.jsx
@@ -1,7 +1,0 @@
-import EditComponent from "../../components/edit-user-component";
-
-function Edit() {
-  return <EditComponent />;
-}
-
-export default Edit;
